reorg commands
